pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "flask>=3.1.0",
    "flask-cors>=5.0.1",
    "numpy>=2.2.3",
    "pillow>=11.1.0",
    "requests>=2.32.3",
    "segment-anything>=1.0",
    "torch>=2.6.0",
    "torchvision>=0.21.0",
    "uvicorn>=0.34.0",
]
